+++ mobs/nakki_v2/NakkiV2.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 
 using Jakojaannos.WisperingWoods.Characters.Player;
 using Jakojaannos.WisperingWoods.Util.Editor;
+using Jakojaannos.WisperingWoods.Util;
 
 namespace Jakojaannos.WisperingWoods;
 
@@ -360,44 +361,5 @@ public partial class NakkiV2 : Path2D {
 
 	private void SinkLilypads() {
 		EmitSignal(SignalName.LilypadAttackSignal);
-	}
-}
-
-/// <summary>
-/// Async extension to AnimationPlayer to allow awaiting on the playing animation.
-/// </summary>
-static class AnimationPlayerAsyncExtension {
-	public static async Task PlayAsync(this AnimationPlayer animationPlayer, string animation) {
-		var completion = new TaskCompletionSource();
-
-		void OnFinished(StringName finished) {
-			if (animation == finished) {
-				completion.SetResult();
-			} else {
-				// Some other animation finished. This is unexpected, cancel the task.
-				completion.SetCanceled();
-			}
-		}
-
-		void OnAnimationChanged(StringName oldAnim, StringName newAnim) {
-			if (newAnim != animation) {
-				// Some other animation started. This is unexpected, cancel the task.
-				completion.SetCanceled();
-			}
-		}
-
-		animationPlayer.AnimationFinished += OnFinished;
-		animationPlayer.AnimationChanged += OnAnimationChanged;
-
-		animationPlayer.Play(animation);
-
-		try {
-			// Wait until the animation finishes or something unexpected happens
-			await completion.Task;
-		} finally {
-			// Clean up the temporary event handlers afterwards
-			animationPlayer.AnimationFinished -= OnFinished;
-			animationPlayer.AnimationChanged -= OnAnimationChanged;
-		}
 	}
 }
